Renamed noncartesian_encoding files
