+++ src/xGadgetron/cGadgetron/cgadgetron.cpp
@@ -41,7 +41,7 @@ limitations under the License.
 // #include "sirf/Gadgetron/FourierEncoding.h"
 
 #if GADGETRON_TOOLBOXES_AVAILABLE
-    #include "sirf/Gadgetron/noncartesian_encoding.h"
+    #include "sirf/Gadgetron/NonCartesianEncoding.h"
 #endif
