+++ src/xGadgetron/cGadgetron/include/sirf/Gadgetron/NonCartesianEncoding.h
